Refactor: insert query
Refactored insert into user_table query to update the name and email if the user_id is already present in the table. This is done according to the information from DX Auth server to maintain user info. This is updated only when the user requests for consumer, provider or delegate specific API